+++ compiler/src/main/java/com/siimkinks/sqlitemagic/writer/ModelPersistingGenerator.java
@@ -31,10 +31,14 @@ import static com.siimkinks.sqlitemagic.WriterUtil.CONTENT_VALUES;
 import static com.siimkinks.sqlitemagic.WriterUtil.LOG_UTIL;
 import static com.siimkinks.sqlitemagic.WriterUtil.OPERATION_FAILED_EXCEPTION;
 import static com.siimkinks.sqlitemagic.WriterUtil.SQLITE_MAGIC;
+import static com.siimkinks.sqlitemagic.WriterUtil.SUBSCRIPTION;
+import static com.siimkinks.sqlitemagic.WriterUtil.SUBSCRIPTIONS;
 import static com.siimkinks.sqlitemagic.WriterUtil.TRANSACTION;
 import static com.siimkinks.sqlitemagic.WriterUtil.addTableTriggersSendingStatement;
 import static com.siimkinks.sqlitemagic.WriterUtil.codeBlockEnd;
 import static com.siimkinks.sqlitemagic.WriterUtil.dbVariableFromPresentConnectionVariable;
+import static com.siimkinks.sqlitemagic.WriterUtil.emitterOnCompleted;
+import static com.siimkinks.sqlitemagic.WriterUtil.emitterOnError;
 import static com.siimkinks.sqlitemagic.WriterUtil.ifNotSubscriberUnsubscribed;
 import static com.siimkinks.sqlitemagic.WriterUtil.subscriberOnCompleted;
 import static com.siimkinks.sqlitemagic.WriterUtil.subscriberOnError;
@@ -47,7 +51,9 @@ import static com.siimkinks.sqlitemagic.util.NameConst.METHOD_BIND_TO_CONTENT_VA
 import static com.siimkinks.sqlitemagic.util.NameConst.METHOD_BIND_TO_NOT_NULL_CONTENT_VALUES;
 import static com.siimkinks.sqlitemagic.util.NameConst.METHOD_SET_ID;
 import static com.siimkinks.sqlitemagic.writer.ModelWriter.DB_CONNECTION_VARIABLE;
+import static com.siimkinks.sqlitemagic.writer.ModelWriter.EMITTER_VARIABLE;
 import static com.siimkinks.sqlitemagic.writer.ModelWriter.ENTITY_VARIABLE;
+import static com.siimkinks.sqlitemagic.writer.ModelWriter.SUBSCRIPTION_VARIABLE;
 import static com.siimkinks.sqlitemagic.writer.ModelWriter.TRANSACTION_VARIABLE;
 
 // FIXME !!! check logging generation
@@ -466,6 +472,17 @@ public class ModelPersistingGenerator implements ModelPartGenerator {
         .beginControlFlow("try");
   }
 
+  public static void addSubscriptionForEmitter(MethodSpec.Builder builder) {
+    builder
+        .addStatement("final $T $L = $T.empty()",
+            SUBSCRIPTION,
+            SUBSCRIPTION_VARIABLE,
+            SUBSCRIPTIONS)
+        .addStatement("$L.setSubscription($L)",
+            EMITTER_VARIABLE,
+            SUBSCRIPTION_VARIABLE);
+  }
+
   static void addTransactionEndBlock(@NonNull MethodSpec.Builder builder, @NonNull Set<TableElement> allTableTriggers,
                                      @NonNull String returnStatement, @NonNull String failReturnStatement) {
     addTransactionEndBlock(builder, allTableTriggers, CodeBlock.builder().addStatement(returnStatement).build(), failReturnStatement);
@@ -524,6 +541,22 @@ public class ModelPersistingGenerator implements ModelPartGenerator {
         .beginControlFlow(ifNotSubscriberUnsubscribed())
         .addStatement(subscriberOnSuccess(successValue))
         .endControlFlow();
+    addTableTriggersSendingStatement(builder, allTableTriggers);
+    builder.endControlFlow()
+        .endControlFlow();
+  }
+
+  public static void addRxCompletableEmitterTransactionEndBlock(@NonNull MethodSpec.Builder builder,
+                                                                @NonNull Set<TableElement> allTableTriggers) {
+    builder
+        .addStatement("$L.markSuccessful()", TRANSACTION_VARIABLE)
+        .addStatement("success = true")
+        .nextControlFlow("catch ($T e)", Throwable.class)
+        .addStatement(emitterOnError())
+        .nextControlFlow("finally")
+        .addStatement("$L.end()", TRANSACTION_VARIABLE)
+        .beginControlFlow("if (success)")
+        .addStatement(emitterOnCompleted());
     addTableTriggersSendingStatement(builder, allTableTriggers);
     builder.endControlFlow()
         .endControlFlow();