+++ runtime/src/main/java/com/siimkinks/sqlitemagic/FunctionColumn.java
@@ -12,13 +12,15 @@ import com.siimkinks.sqlitemagic.internal.StringArraySet;
 import java.util.ArrayList;
 import java.util.LinkedList;
 
-final class FunctionColumn<
-    T, // exact type
-    R, // return type (when this column is queried)
-    ET, // equivalent type
-    P>  // parent table type
-    extends NumericColumn<T, R, ET, P> {
-
+/**
+ * A column used in queries and conditions.
+ *
+ * @param <T>  Exact type
+ * @param <R>  Return type (when this column is queried)
+ * @param <ET> Equivalent type
+ * @param <P>  Parent table type
+ */
+final class FunctionColumn<T, R, ET, P> extends NumericColumn<T, R, ET, P> {
   @NonNull
   private final Column[] wrappedColumns;
   @NonNull
@@ -27,6 +29,7 @@ final class FunctionColumn<
   private final String separator;
   @NonNull
   private final String suffix;
+  private boolean compiledToSelection = false;
 
   FunctionColumn(@NonNull Table<P> table,
                  @NonNull @Size(min = 1) Column[] wrappedColumns,
@@ -56,6 +59,10 @@ final class FunctionColumn<
 
   @Override
   void appendSql(@NonNull StringBuilder sb) {
+    if (compiledToSelection && hasAlias()) {
+      sb.append(getAppendableAlias());
+      return;
+    }
     sb.append(prefix);
     final Column[] wrappedColumns = this.wrappedColumns;
     final int length = this.wrappedColumns.length;
@@ -71,6 +78,10 @@ final class FunctionColumn<
 
   @Override
   void appendSql(@NonNull StringBuilder sb, @NonNull SimpleArrayMap<String, LinkedList<String>> systemRenamedTables) {
+    if (compiledToSelection && hasAlias()) {
+      sb.append(getAppendableAlias());
+      return;
+    }
     sb.append(prefix);
     final Column[] wrappedColumns = this.wrappedColumns;
     final int length = this.wrappedColumns.length;
@@ -116,6 +127,7 @@ final class FunctionColumn<
     appendSql(compiledCols);
     appendAliasDeclarationIfNeeded(compiledCols);
     putColumnPosition(columnPositions, null, columnOffset, this);
+    compiledToSelection = true;
     return columnOffset + 1;
   }
 
@@ -124,6 +136,7 @@ final class FunctionColumn<
     appendSql(compiledCols, systemRenamedTables);
     appendAliasDeclarationIfNeeded(compiledCols);
     putColumnPosition(columnPositions, null, columnOffset, this);
+    compiledToSelection = true;
     return columnOffset + 1;
   }