+++ runtime/src/main/java/com/siimkinks/sqlitemagic/entity/EntityBulkUpdateBuilder.java
@@ -3,14 +3,13 @@ package com.siimkinks.sqlitemagic.entity;
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 
-import rx.Single;
+import rx.Completable;
 
 /**
  * Builder for bulk update operation.
  *
- * @param <T> Operation target type
  */
-public interface EntityBulkUpdateBuilder<T> extends ConnectionProvidedOperation<EntityBulkUpdateBuilder<T>> {
+public interface EntityBulkUpdateBuilder extends ConnectionProvidedOperation<EntityBulkUpdateBuilder> {
   /**
    * Execute this configured bulk update operation against a database.
    * Operation will be executed inside a transaction.
@@ -21,16 +20,16 @@ public interface EntityBulkUpdateBuilder<T> extends ConnectionProvidedOperation<
   boolean execute();
 
   /**
-   * Creates a {@link Single} that when subscribed to executes this configured bulk
-   * update operation against a database and emits operation result to downstream
-   * only once. Operation will be executed inside a transaction. If the operation was
-   * successful then {@code true} will be emitted to downstream.
+   * Creates a {@link Completable} that when subscribed to executes this configured bulk
+   * update operation against a database and emits operation result to downstream.
+   * Operation will be executed inside a transaction. If the operation was
+   * successful then complete will be emitted to downstream.
    * If the operation failed then it will be rolled back and error will be emitted to downstream.
    *
-   * @return Deferred {@link Single} that when subscribed to executes the operation and emits
+   * @return Deferred {@link Completable} that when subscribed to executes the operation and emits
    * its result to downstream
    */
   @NonNull
   @CheckResult
-  Single<Boolean> observe();
+  Completable observe();
 }