+++ runtime/src/main/java/com/siimkinks/sqlitemagic/entity/EntityUpdateBuilder.java
@@ -5,7 +5,7 @@ import android.support.annotation.NonNull;
 
 import com.siimkinks.sqlitemagic.ConflictAlgorithm;
 
-import rx.Single;
+import rx.Completable;
 
 /**
  * Builder for update operation.
@@ -33,17 +33,17 @@ public interface EntityUpdateBuilder extends ConnectionProvidedOperation<EntityU
   boolean execute();
 
   /**
-   * Creates a {@link Single} that when subscribed to executes this configured
-   * update operation against a database and emits the operation result to downstream
-   * only once. Operation will be executed inside a transaction if the updated entity has
+   * Creates a {@link Completable} that when subscribed to executes this configured
+   * update operation against a database and emits the operation result to downstream.
+   * Operation will be executed inside a transaction if the updated entity has
    * complex columns which also need to be updated.
-   * If the operation was successful then {@code true} will be emitted to downstream.
+   * If the operation was successful then complete will be emitted to downstream.
    * If the operation failed then it will be rolled back and error will be emitted to downstream.
    *
-   * @return Deferred {@link Single} that when subscribed to executes the operation and emits
+   * @return Deferred {@link Completable} that when subscribed to executes the operation and emits
    * its result to downstream
    */
   @NonNull
   @CheckResult
-  Single<Boolean> observe();
+  Completable observe();
 }